requirements.txt
phidata
python-dotenv
fastapi[standard]
uvicorn[standard]
sqlalchemy
duckduckgo-search
openai
yfinance